More adjustments and fixes
